Fixed ground truth rot mat loading and added overwrite checks for re-writing data

+++ src/vtr_testing_radar/script/test_localization.sh
@@ -12,6 +12,19 @@ echo "Running localization on sequence ${LOC_INPUT} to reference sequence ${ODO_
 # Source the VTR environment with the testing package
 source ${VTRRROOT}/install/setup.bash
 
+graph_dir=${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}/graph
+if [ -d $graph_dir ]; then
+  # Count the number of directories inside "graph"
+  dir_count=$(ls -l $graph_dir | grep -c ^d)
+  
+  if [ $dir_count -gt 1 ]; then
+    read -p "The directory $graph_dir is not empty and contains $dir_count other directories. Do you want to delete it and create an empty one? (yes/no) " response
+    if [ "$response" == "no" ]; then
+      exit
+    fi
+  fi
+fi
+
 rm -r ${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}
 mkdir -p ${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}
 cp -r ${VTRRRESULT}/${ODO_INPUT}/${ODO_INPUT}/* ${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}
